+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
- - Allow scatter-gather injection via `DependencyMap::insert_container` ([PR #7](https://github.com/p0lunin/dptree/pull/7)).
+ - `DependencyMap::insert_container` ([PR #7](https://github.com/p0lunin/dptree/pull/7)).
 
 ## 0.1.1 - 2022-03-21